FEAT: ManagementButton moves to local widget to global widget
