+++ lib/src/features/friend_management/presentation/pages/friend_management_page/local_widgets/add_friend_tile.dart
@@ -1,4 +1,5 @@
 import 'package:chat_application/src/core/routes/app_routes.dart';
+import 'package:chat_application/src/core/widgets/management_button.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
@@ -7,9 +8,8 @@ class AddFriendTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ListTile(
-      title: const Text('친구 추가'),
-      trailing: const Icon(Icons.chevron_right),
+    return ManagementButton(
+      name: '친구 추가',
       onTap: () => context.pushNamed(FriendManagementRoutes.addFriend.name),
     );
   }
