+++ lib/src/features/settings/presentation/pages/settings_page/local_widgets/account_management_button.dart
@@ -1,5 +1,5 @@
 import 'package:chat_application/src/core/routes/app_routes.dart';
-import 'package:chat_application/src/features/settings/presentation/pages/settings_page/local_widgets/management_button.dart';
+import 'package:chat_application/src/core/widgets/management_button.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
